requirements.txt
alabaster==0.7.12
Babel==2.8.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==6.7
colorama==0.4.3
cryptography==2.8
cssselect2==0.2.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-cors-headers==3.3.0
django-filter==2.2.0
django-import-export==2.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-simple-email-confirmation==0.70
django-wkhtmltopdf==3.3.0
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.16
et-xmlfile==1.0.1
feedparser==5.2.1
fpdf==1.7.2
#get==2019.4.13
#gitsome==0.8.4
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
imagesize==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jwcrypto==0.7
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.19.1
numpydoc==0.9.2
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
packaging==20.4
pandas==1.1.1
Pillow==6.2.1
ply==3.11
#post==2019.4.13
prompt-toolkit==2.0.10
public==2019.4.13
pycparser==2.19
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
Pyphen==0.9.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.2
query-string==2019.4.13
reportlab==3.5.49
#request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
snowballstemmer==2.0.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
tablib==0.14.0
tinycss2==1.0.2
twilio==6.34.0
unicodecsv==0.14.1
uritemplate==3.0.1
uritemplate.py==3.0.2
urllib3==1.25.7
wcwidth==0.1.9
WeasyPrint==50
webencodings==0.5.1
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
